[nfc] Disentangle and Modularize JSG
- Factor several files out of the main JSG target
- Significantly reduce header interdependencies/cycles
- Start splitting up uses of <v8.h> into constituent headers.

This will make future refactoring of JSG easier.
