+++ src/workerd/io/BUILD.bazel
@@ -124,6 +124,8 @@ wd_cc_library(
         "//src/workerd/api:url",
         "//src/workerd/api:urlpattern",
         "//src/workerd/jsg",
+        "//src/workerd/jsg:inspector",
+        "//src/workerd/jsg:script",
         "//src/workerd/util:autogate",
         "//src/workerd/util:completion-membrane",
         "//src/workerd/util:exception",
