+++ src/workerd/jsg/iterator.h
@@ -4,10 +4,9 @@
 
 #pragma once
 
-#include "jsg.h"
-#include "struct.h"
-
+#include <workerd/jsg/jsg.h>
 #include <workerd/jsg/memory.h>
+#include <workerd/jsg/struct.h>
 
 #include <concepts>
 #include <deque>
