+++ src/workerd/jsg/multiple-typewrappers-test.c++
@@ -1,7 +1,7 @@
-#include <workerd/io/compatibility-date.h>
-#include <workerd/io/observer.h>
+#include <workerd/io/compatibility-date.capnp.h>
 #include <workerd/jsg/jsg.h>
 #include <workerd/jsg/setup.h>
+#include <workerd/jsg/type-wrapper.h>
 
 #include <capnp/message.h>
 #include <kj/test.h>
