+++ src/workerd/jsg/promise.h
@@ -8,6 +8,8 @@
 #include "util.h"
 #include "wrappable.h"
 
+#include <v8-promise.h>
+
 #include <kj/async.h>
 #include <kj/table.h>
