+++ src/workerd/jsg/resource.h
@@ -10,12 +10,15 @@
 // can call back to the class's methods. This differs from, say, a struct type, which will be deeply
 // converted into a JS object when passed into JS.
 
-#include "meta.h"
+#include "ser.h"
 #include "util.h"
 #include "wrappable.h"
 
 #include <workerd/jsg/memory.h>
+#include <workerd/jsg/meta.h>
 #include <workerd/jsg/modules.capnp.h>
+
+#include <v8-template.h>
 
 #include <kj/debug.h>
 #include <kj/map.h>
